Move selfavatar tests to blob.rs
